// src/app/interfaces/microServices.ts
export interface IFormatoAuditoria {
    nU_FORMATOCD_AUDCDHC: number;
    nU_IDHISTCLINIC_AUDCDHC: number;
    cL_HISTORIACLINICA_AUDCDHC: string;
    tX_PERIODO_AUDCDHC: string;
    fE_FECHAENVIO_AUDCDHC: Date|string;
    nU_IDPROFMED_AUDCDHC: number;
    tX_PROFMED_AUDCDHC: string;
    nU_IDAFILIADO_AFIL: number;
    tX_IDENTIFICACION_AFIL: string;
    tX_NOMBREAFIL_AUDCDHC: string;
}